+++ source/Datagrams.NET/BinaryHelper.cs
@@ -1,5 +1,5 @@
-﻿using DatagramsNet.Datagrams.NET.Logger;
-using DatagramsNet.Datagrams.NET.Prefixes;
+﻿using DatagramsNet.Logging;
+using DatagramsNet.Prefixes;
 using System.Runtime.InteropServices;
 
 namespace DatagramsNet
