+++ source/Datagrams.NET/Logging/Reading/CommandExecuting/CommandFunctionHolder.cs
@@ -1,9 +1,9 @@
-﻿using DatagramsNet.Datagrams.NET.Logger.Reader.Interfaces;
+﻿using DatagramsNet.Logging.Reading.Interfaces;
 using System.Reflection;
 
-namespace DatagramsNet.Datagrams.NET.Logger.Reader.CommandExecuting
+namespace DatagramsNet.Logging.Reading.CommandExecuting
 {
-    internal struct ActionCommand 
+    internal struct ActionCommand
     {
         public Type CommandType { get; set; }
 
@@ -14,10 +14,10 @@ namespace DatagramsNet.Datagrams.NET.Logger.Reader.CommandExecuting
     {
         private static List<ActionCommand> actionCommands = new();
 
-        public static ICommandAction GetFunction<T>(T command)  where T : ICommandAction
+        public static ICommandAction GetFunction<T>(T command) where T : ICommandAction
         {
             var firstCommand = actionCommands.FirstOrDefault(n => command.GetType() == n.CommandType);
-            if (firstCommand.CommandType is not null) 
+            if (firstCommand.CommandType is not null)
             {
                 command.CommandAction = firstCommand.CommandAction;
                 return command;
@@ -43,13 +43,13 @@ namespace DatagramsNet.Datagrams.NET.Logger.Reader.CommandExecuting
                         methodAction = commandActions[0];
                 }
                 //else
-                    //throw new Exception("Multiple actions for one command are not supported");
+                //throw new Exception("Multiple actions for one command are not supported");
             }
             command.CommandAction = methodAction;
             return command;
         }
 
-        private static IEnumerable<Action> GetCommandActions<T>(Type assembly, T command)  where T : ICommandAction
+        private static IEnumerable<Action> GetCommandActions<T>(Type assembly, T command) where T : ICommandAction
         {
             var assemblyMethods = assembly.GetMethods().Where(n => n.GetCustomAttributes(typeof(CommandFunctionAttribute<T>), true).Length > 0).ToArray();
             for (int i = 0; i < assemblyMethods.Length; i++)
@@ -57,7 +57,7 @@ namespace DatagramsNet.Datagrams.NET.Logger.Reader.CommandExecuting
                 var commandAttribute = (CommandFunctionAttribute<T>[])assemblyMethods[i].GetCustomAttributes(typeof(CommandFunctionAttribute<T>), true);
                 for (int j = 0; j < commandAttribute.Length; j++)
                 {
-                    if (((commandAttribute[j].Command.GetType() == command.GetType()))) 
+                    if (commandAttribute[j].Command.GetType() == command.GetType())
                     {
                         var delegateAction = (Action)Delegate.CreateDelegate(typeof(Action), assemblyMethods[i]);
                         yield return delegateAction;
@@ -65,5 +65,5 @@ namespace DatagramsNet.Datagrams.NET.Logger.Reader.CommandExecuting
                 }
             }
         }
-    } 
+    }
 }
