+++ source/Datagrams.NET/Logging/Reading/Indexes/ArgumentIndex.cs
@@ -1,7 +1,7 @@
-﻿using DatagramsNet.Datagrams.NET.Logger.Reader.Interfaces;
-using DatagramsNet.Datagrams.NET.Prefixes;
+﻿using DatagramsNet.Prefixes;
+using DatagramsNet.Logging.Reading.Interfaces;
 
-namespace DatagramsNet.Datagrams.NET.Logger.Reader.Indexes
+namespace DatagramsNet.Logging.Reading.Indexes
 {
     internal sealed class ArgumentIndex : IIndex<Argument[], ArgumentIndex>
     {
@@ -11,18 +11,18 @@ namespace DatagramsNet.Datagrams.NET.Logger.Reader.Indexes
 
         public Argument[] Value { get; set; }
 
-        public ArgumentIndex GetIndex(string command, char separator, int index) 
+        public ArgumentIndex GetIndex(string command, char separator, int index)
         {
             var values = command.Split(separator);
-            var indexValue = (values.Length - 1) >= (index + 1) ? values[index + 1] : null;
+            var indexValue = values.Length - 1 >= index + 1 ? values[index + 1] : null;
             List<Argument> properArguments = new List<Argument>();
             if (indexValue is not null)
                 properArguments = GetProperArguments(Command.Arguments, indexValue.ToCharArray()).ToList();
 
-            return new ArgumentIndex() { Command = this.Command, Value = properArguments.ToArray()};
+            return new ArgumentIndex() { Command = Command, Value = properArguments.ToArray() };
         }
 
-        private IEnumerable<Argument> GetProperArguments(Argument[] baseArguments, char[] argumentsChar) 
+        private IEnumerable<Argument> GetProperArguments(Argument[] baseArguments, char[] argumentsChar)
         {
             for (int i = 0; i < argumentsChar.Length; i++)
             {
@@ -32,10 +32,10 @@ namespace DatagramsNet.Datagrams.NET.Logger.Reader.Indexes
                     if (argumentsChar[i] == baseArguments[j].Character)
                     {
                         index++;
-                        yield return new Argument() {Character = argumentsChar[i] };
+                        yield return new Argument() { Character = argumentsChar[i] };
                     }
                 }
-                if(index == 0)
+                if (index == 0)
                     ServerLogger.Log<ErrorPrefix>($"Sorry but argument '{argumentsChar[i]}' was not found");
             }
         }
