+++ source/Datagrams.NET/Logging/Reading/Interfaces/ICommandAction.cs
@@ -1,5 +1,4 @@
-﻿
-namespace DatagramsNet.Datagrams.NET.Logger.Reader.Interfaces
+﻿namespace DatagramsNet.Logging.Reading.Interfaces
 {
     public interface ICommandAction
     {
