+++ source/Datagrams.NET/Prefixes/ErrorPrefix.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace DatagramsNet.Datagrams.NET.Prefixes
+namespace DatagramsNet.Prefixes
 {
     public sealed class ErrorPrefix : StandardPrefix
     {
