+++ source/Datagrams.NET/Prefixes/WarningPrefix.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace DatagramsNet.Datagrams.NET.Prefixes
+namespace DatagramsNet.Prefixes
 {
     public sealed class WarningPrefix : StandardPrefix
     {
